async_log: decide the log format and implement it
Now the log is outputed to stdout directly, this behavior
will be changed in the future.

+++ read_conf.h
@@ -7,7 +7,7 @@
 struct muQconf
 {
     uint16_t v4port;
-    size_t max_thread_num;
+    int max_thread_num;
     int log_level;
 };
